+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/conv10_50/ValueSet10_50.java
@@ -1,8 +1,9 @@
 package org.hl7.fhir.convertors.conv10_50;
 
 import java.util.List;
-import org.hl7.fhir.convertors.advisors.VersionConvertorAdvisor50;
+
 import org.hl7.fhir.convertors.VersionConvertor_10_50;
+import org.hl7.fhir.convertors.advisors.VersionConvertorAdvisor50;
 import org.hl7.fhir.dstu2.model.ValueSet;
 import org.hl7.fhir.exceptions.FHIRException;
 import org.hl7.fhir.r5.model.BooleanType;