+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/misc/IGR2ConvertorAdvisor.java
@@ -1,7 +1,5 @@
 package org.hl7.fhir.convertors.misc;
 
-import org.hl7.fhir.convertors.advisors.VersionConvertorAdvisor40;
-
 /*
   Copyright (c) 2011+, HL7, Inc.
   All rights reserved.
@@ -33,16 +31,16 @@ import org.hl7.fhir.convertors.advisors.VersionConvertorAdvisor40;
 
 
 
-import org.hl7.fhir.dstu2.model.Resource;
-import org.hl7.fhir.exceptions.FHIRException;
+import org.hl7.fhir.convertors.advisors.VersionConvertorAdvisor40;
 import org.hl7.fhir.r4.model.Bundle.BundleEntryComponent;
 import org.hl7.fhir.r4.model.CodeSystem;
 import org.hl7.fhir.r4.model.ValueSet;
+import org.hl7.fhir.r5.model.FhirPublication;
 
 public class IGR2ConvertorAdvisor implements VersionConvertorAdvisor40 {
 
   @Override
-  public boolean ignoreEntry(BundleEntryComponent src) {
+  public boolean ignoreEntry(BundleEntryComponent src, FhirPublication publication) {
     return false;
   }