+++ RELEASE_NOTES.md
@@ -6,3 +6,4 @@
 * Issue 484 https://github.com/hapifhir/org.hl7.fhir.core/issues/484
 * Adding Kotlin to the build process
 * random cleaning up in convertors
+* Adding version option to ignore bundle entry in conversion advisors